+++ agent/cmd/service_windows.go
@@ -3,77 +3,59 @@ package cmd
 import (
 	"context"
 	"errors"
-	"fmt"
 	"os"
-	"sync"
-
+	"os/signal"
 	runtimedebug "runtime/debug"
+	"sync"
+	"syscall"
 
 	"github.com/sensu/sensu-go/agent"
-	"github.com/sirupsen/logrus"
 	"golang.org/x/sys/windows/svc"
-	"golang.org/x/sys/windows/svc/debug"
-	"golang.org/x/sys/windows/svc/eventlog"
 )
 
 var (
-	_            svc.Handler = &Service{}
-	elog         debug.Log
 	AgentNewFunc = agent.NewAgentContext
 )
 
-func NewService(args []string) *Service {
-	return &Service{args: args}
+func NewService(cfg *agent.Config) *Service {
+	return &Service{cfg: cfg}
 }
 
 type Service struct {
-	args []string
-	wg   sync.WaitGroup
-	mu   sync.Mutex
+	cfg *agent.Config
+	wg  sync.WaitGroup
 }
 
-func (s *Service) start(ctx context.Context, args []string, changes chan<- svc.Status) chan error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.wg.Wait()
+func (s *Service) start(ctx context.Context, cancel context.CancelFunc, changes chan<- svc.Status) chan error {
 	s.wg.Add(1)
 	result := make(chan error, 1)
+	defer func() {
+		if e := recover(); e != nil {
+			changes <- svc.Status{State: svc.Stopped}
+			stack := runtimedebug.Stack()
+			result <- errors.New(string(stack))
+		}
+	}()
+	changes <- svc.Status{State: svc.StartPending}
+	accepts := svc.AcceptShutdown | svc.AcceptStop
+	changes <- svc.Status{State: svc.Running, Accepts: accepts}
+
+	sensuAgent, err := agent.NewAgentContext(ctx, s.cfg)
+	if err != nil {
+		result <- err
+		return result
+	}
+
+	sigs := make(chan os.Signal, 1)
+	signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM)
 	go func() {
-		defer func() {
-			if e := recover(); e != nil {
-				changes <- svc.Status{State: svc.Stopped}
-				stack := runtimedebug.Stack()
-				result <- errors.New(string(stack))
-			}
-		}()
+		defer cancel()
+		logger.Info("signal received: ", <-sigs)
+	}()
+
+	go func() {
 		defer s.wg.Done()
-		changes <- svc.Status{State: svc.StartPending}
-		// Start service here
-		binPath, err := exePath()
-		if err != nil {
-			panic(err)
-		}
-		configFile := args[0]
-		logPath := args[1]
-		logFile, err := os.OpenFile(logPath, os.O_CREATE|os.O_APPEND|os.O_WRONLY, 0600)
-		if err != nil {
-			result <- fmt.Errorf("service quit: cant't open log file: %s", err)
-		}
-		defer logFile.Close()
-
-		logrus.SetFormatter(&logrus.JSONFormatter{})
-		logrus.SetOutput(logFile)
-		logger := logrus.WithFields(logrus.Fields{
-			"component": "cmd",
-		})
-
-		args = []string{binPath, "start", "-c", configFile}
-		command := StartCommand(AgentNewFunc)
-		accepts := svc.AcceptShutdown | svc.AcceptStop
-		changes <- svc.Status{State: svc.Running, Accepts: accepts}
-
-		if err := command.Execute(); err != nil {
-			logger.WithError(err).Error("sensu-agent exited with error")
+		if err := sensuAgent.Run(ctx); err != nil {
 			result <- err
 		}
 	}()
@@ -82,15 +64,14 @@ func (s *Service) start(ctx context.Context, args []string, changes chan<- svc.S
 
 func (s *Service) Execute(_ []string, r <-chan svc.ChangeRequest, changes chan<- svc.Status) (bool, uint32) {
 	ctx, cancel := context.WithCancel(context.Background())
-	errs := s.start(ctx, s.args, changes)
-	elog, _ := eventlog.Open(serviceName)
-	defer elog.Close()
+	logger.Info("sensu-agent service starting")
+	errs := s.start(ctx, cancel, changes)
 	for {
 		select {
 		case req := <-r:
 			switch req.Cmd {
 			case svc.Stop, svc.Shutdown:
-				elog.Info(1, "service shutting down")
+				logger.Info("sensu-agent shutting down")
 				changes <- svc.Status{State: svc.StopPending}
 				cancel()
 				s.wg.Wait()
@@ -98,23 +79,9 @@ func (s *Service) Execute(_ []string, r <-chan svc.ChangeRequest, changes chan<-
 				return false, 0
 			}
 		case err := <-errs:
-			elog.Error(1, fmt.Sprintf("restarting due to error (%v) %s", s.args, err))
-			s.start(ctx, s.args, changes)
+			logger.WithError(err).Error("fatal error")
+			return false, 1
 		}
 	}
 	return false, 0
-}
-
-func runService(args []string) error {
-	elog, err := eventlog.Open(serviceName)
-	if err != nil {
-		return err
-	}
-	defer elog.Close()
-	elog.Info(1, fmt.Sprintf("starting %s service (%v)", serviceName, args))
-	if err := svc.Run(serviceName, NewService(args)); err != nil {
-		return err
-	}
-	elog.Info(1, fmt.Sprintf("%s service terminated", serviceName))
-	return nil
 }